% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cbsa_code.R
\name{find_cbsa_code}
\alias{find_cbsa_code}
\title{This function returns cbsa code by matching fuzzy metro names}
\usage{
find_cbsa_code(msa)
}
\arguments{
\item{msa}{name of a metro}
}
\value{
5 digit CBSA code
}
\description{
This function returns cbsa code by matching fuzzy metro names
}
