% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cbsa_counties.R
\name{find_cbsa_counties}
\alias{find_cbsa_counties}
\title{find cbsa_code by short name and return all counties within the CBSA}
\usage{
find_cbsa_counties(msa)
}
\arguments{
\item{msa}{name of a metro}
}
\value{
dataframe of counties with matching CBSA
}
\description{
find cbsa_code by short name and return all counties within the CBSA
}
