% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl_online.R
\name{readxl_online}
\alias{readxl_online}
\title{Read excel files from online locations}
\usage{
readxl_online(url, type = NULL, ...)
}
\arguments{
\item{url}{Path to the file}

\item{type}{.xlsx, .xls or .zip}
}
\value{
A dataframe
}
\description{
This function loads xls or zipped xls file from online locations. I
}
